+++ HISTORY.md
@@ -1,6 +1,7 @@
 # History
 
 ---
+
 ## 1.0.0
 
 `tag:fixed` [#20](https://github.com/aralejs/switchable/issues/20) Circular bugfix.